pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "signalops"
version = "0.1.0"
description = "Agentic social lead finder + outreach workbench"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "httpx>=0.27",
    "anthropic>=0.40",
    "openai>=1.50",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.11",
    "ruff>=0.6",
    "respx>=0.21",
    "types-PyYAML>=6.0",
]

[project.scripts]
signalops = "signalops.cli.main:cli"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
strict = true

[[tool.mypy.overrides]]
module = [
    "sklearn.*",
    "joblib",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
